+++ web/components/Account/LearningDistribution/types.ts
@@ -1,0 +1,6 @@
+import { BarDataType } from '@/components/UI/PercentageBar/types'
+
+export type Props = {
+  barData: BarDataType[]
+  description?: string
+}
