+++ web/pages/account/learning/index.tsx
@@ -1,6 +1,6 @@
 import AccountLayout from '@/components/Layout/AccountLayout/AccountLayout'
 import LearningGoalBar from '@/components/Account/LearningGoalBar/LearningGoalBar'
-import { Box, Grid, styled, Typography, useTheme } from '@mui/material'
+import { Grid, Typography } from '@mui/material'
 import Button from '@/components/UI/Button/Button'
 import useAuth from '@/hooks/useAuth'
 import { useQuery } from 'react-query'
@@ -10,33 +10,12 @@ import Tabs from '@/components/UI/Tabs/Tabs'
 import LearningRecordList from '@/components/Account/LearningRecordList/LearningRecordList'
 import { fetchMeLearning } from '@/service/me'
 import AccountCard from '@/components/UI/Cards/AccountCard/AccountCard'
-import Tooltip from '@/components/UI/Tooltip/Tooltip'
 import { useMemo } from 'react'
 import { BarDataType } from '@/components/UI/PercentageBar/types'
-import PercentageBar from '@/components/UI/PercentageBar/PercentageBar'
-
-const GraphDescription = styled(Typography)`
-  display: inline-flex;
-  margin-bottom: ${(p) => p.theme.spacing(2)};
-  position: relative;
-  min-height: 24px;
-  > span {
-    display: inline-flex;
-    align-items: center;
-    margin-right: ${(p) => p.theme.spacing(2)};
-  }
-  .dot {
-    border: 2px solid ${(p) => p.theme.colors.black};
-    width: 20px;
-    height: 20px;
-    border-radius: 50%;
-    margin-right: ${(p) => p.theme.spacing(1)};
-  }
-`
+import LearningDistribution from '@/components/Account/LearningDistribution/LearningDistribution'
 
 const LearningPage = () => {
   const { authFetch } = useAuth()
-  const { colors } = useTheme()
 
   const { data } = useQuery<MeLearningRecord[]>(
     Query.MeLearning,
@@ -57,9 +36,13 @@ const LearningPage = () => {
     const total = totalWork + totalSocial + totalFormal
 
     return [
-      { label: 'On the job', percentage: (totalWork / total) * 100, color: 'blue1' },
-      { label: 'Social', percentage: (totalSocial / total) * 100, color: 'grey2' },
-      { label: 'Formal', percentage: (totalFormal / total) * 100, color: 'black' }
+      {
+        label: 'On the job',
+        percentage: (totalWork / total) * 100,
+        color: null
+      },
+      { label: 'Social', percentage: (totalSocial / total) * 100, color: null },
+      { label: 'Formal', percentage: (totalFormal / total) * 100, color: null }
     ]
   }, [data])
 
@@ -81,51 +64,7 @@ const LearningPage = () => {
       <Grid container spacing={5} sx={{ mb: 5 }}>
         <Grid item xs={6}>
           <AccountCard sx={{ height: '100%' }}>
-            <Typography variant="h2" component="h3" gutterBottom>
-              Learning distribution{' '}
-              <Tooltip
-                sx={{ p: 0 }}
-                brandColor="brandLearning"
-                title="To get the most out of your learning you should aim for 70% learning
-                on the job, 20% social and 10% formal training"
-              />
-            </Typography>
-            <GraphDescription variant="body2">
-              {barData.map(({ label, color, percentage }) => (
-                <span key={label}>
-                  <Box
-                    className="dot"
-                    sx={{ backgroundColor: colors[color] }}
-                    component="span"
-                  />{' '}
-                  {label} ({(percentage || 0).toFixed()}%)
-                </span>
-              ))}
-              <Tooltip
-                brandColor="brandLearning"
-                sx={{ p: 0 }}
-                title={
-                  <>
-                    <Typography variant="body2">On the job</Typography> Self-taught
-                    learning by doing, for example reading policies and guidance, using
-                    tools and software to do your job
-                    <Typography variant="body2">Social</Typography>
-                    Learning from colleagues, job shadowing, mentoring, coaching, networks
-                    and communities
-                    <Typography variant="body2">Formal</Typography>
-                    Completing a course on Civil Service Learning, external training,
-                    professional qualifications
-                  </>
-                }
-              />
-            </GraphDescription>
-            <PercentageBar
-              data={barData}
-              marks={[0, 25, 50, 75, 100].map((value) => ({
-                value,
-                label: value.toString()
-              }))}
-            />
+            <LearningDistribution barData={barData} />
           </AccountCard>
         </Grid>
         <Grid item xs={6}>
@@ -133,9 +72,9 @@ const LearningPage = () => {
             <Typography variant="h2" component="h3" gutterBottom>
               Learning goal
             </Typography>
-            <GraphDescription variant="body2" gutterBottom>
+            <Typography variant="body2" gutterBottom>
               You should aim to complete 10 days learning each year
-            </GraphDescription>
+            </Typography>
 
             <LearningGoalBar />
           </AccountCard>
